+++ ruitk/themes/fallen/input.js
@@ -94,11 +94,12 @@ let elements = [
                 if (Array.isArray(box) === false) {
                     box = [box];
                 }
-                for (let item of box) {
-                    if (typeof item === "string") {
-                        form.innerHTML += item; 
-                    } else if (item instanceof HTMLElement) {
-                        form.appendChild(item); 
+                for (let item of info.content) {
+                    if (typeof item === "string" || item instanceof HTMLElement) {
+                        e.append(item); 
+                    } else {
+                        console.warn(`${element.name} Element: item in info.content is not str or HTML element. dumping item to debug`);
+                        console.debug(JSON.parse(JSON.stringify(item)));
                     }
                 }
             }
@@ -212,11 +213,12 @@ let elements = [
             if (Array.isArray(info.question) === false) {
                 info.question = [info.question];
             }
-            for (let item of info.question) {
-                if (typeof item === "string") {
-                    form.innerHTML += item; 
-                } else if (item instanceof HTMLElement) {
-                    form.appendChild(item); 
+            for (let item of info.content) {
+                if (typeof item === "string" || item instanceof HTMLElement) {
+                    e.append(item); 
+                } else {
+                    console.warn(`${element.name} Element: item in info.content is not str or HTML element. dumping item to debug`);
+                    console.debug(JSON.parse(JSON.stringify(item)));
                 }
             }
             return form;